Specified sizes of many ints

+++ src/AOSongAM2315.cpp
@@ -25,7 +25,7 @@
 
 
 // The constructor - because this is I2C, only need the power pin
-AOSongAM2315::AOSongAM2315(int powerPin, int measurementsToAverage)
+AOSongAM2315::AOSongAM2315(int8_t powerPin, uint8_t measurementsToAverage)
     : Sensor(F("AOSongAM2315"), AM2315_NUM_VARIABLES,
              AM2315_WARM_UP_TIME_MS, AM2315_STABILIZATION_TIME_MS, AM2315_MEASUREMENT_TIME_MS,
              powerPin, -1, measurementsToAverage)
