+++ src/MaximDS18.h
@@ -46,8 +46,8 @@
 class MaximDS18 : public Sensor
 {
 public:
-    MaximDS18(DeviceAddress OneWireAddress, int powerPin, int dataPin, int measurementsToAverage = 1);
-    MaximDS18(int powerPin, int dataPin, int measurementsToAverage = 1);
+    MaximDS18(DeviceAddress OneWireAddress, int8_t powerPin, int8_t dataPin, uint8_t measurementsToAverage = 1);
+    MaximDS18(int8_t powerPin, int8_t dataPin, uint8_t measurementsToAverage = 1);
 
     SENSOR_STATUS setup(void) override;
     String getSensorLocation(void) override;
